+++ src/components/Header.jsx
@@ -1,10 +1,12 @@
 import React from 'react';
+import Navigation from './Navigation';
 import './Header.scss';
 
 const Header = () => {
   return (
     <header>
       <h1><a href="/">cook<b>book</b></a></h1>
+      <Navigation />
     </header>
   )
 }