+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.4.0] - 2024-12-23
+
+### Added
+- Mobile navigation
+- Responsive styling
+- Favorites
+- Screenshots folder
+
+### Changed
+- Recipe Layout: moved YT embed below the instructions
+
+### Fixed
+- Invalid Meal ID/no meals found error-handling
+
 ## [0.3.0] - 2024-12-23
 
 ### Added